G7_2022_exam-main/pyproject.toml
[tool.poetry]
name = "2022-exam-monitoring"
version = "0.1.0"
description = ""
authors = ["lili <lilij2079@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
Jinja2 = "^3.0.3"
SQLAlchemy = "^1.4.31"
psycopg2-binary = "^2.9.3"
alembic = "^1.7.6"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
